Adds a One Page all-up archive html file
Adds a One Page all-up archive html file.

It's roughly 70mb of html currently per 10k messages.

This is in very rough shape at the moment.  The formatting is awful, the messages need to be html unencoded and the unix timestamps need to be converted to human readable dates. 

That said, I'm releasing it in this form right now in case someone wants to search for http:// or ftp links to files/locations linked to from a yahoo group.